mcp_servers/window_capture/requirements.txt
Pillow>=9.5.0
